12 Financial Stock llama Index/requirements.txt
openai
ib_insync
pandas
llama-index
jinja2
streamlit
python-dotenv